// packages/core/src/types/types.ts
export interface Link<T extends string> {
  sys: {
    type: 'Link';
    linkType: T;
    id: string;
  };
}

export type ComponentDefinitionVariableType =
  | 'Text'
  | 'RichText'
  | 'Number'
  | 'Date'
  | 'Boolean'
  | 'Location'
  | 'Object';
// | 'Link'
// | 'Array'
// export type ComponentDefinitionVariableArrayItemType = 'Link' | 'Symbol' | 'Component'

export type VariableFormats = 'URL'; // | alphaNum | base64 | email | ipAddress

export type ValidationOption<T extends ComponentDefinitionVariableType> = {
  value: T extends 'Text' ? string : T extends 'Number' ? number : never;
  displayName?: string;
};

export type ComponentDefinitionVariableValidation<T extends ComponentDefinitionVariableType> = {
  required?: boolean;
  in?: ValidationOption<T>[];
  format?: VariableFormats;
};

export interface ComponentDefinitionVariableBase<T extends ComponentDefinitionVariableType> {
  type: T;
  validations?: ComponentDefinitionVariableValidation<T>;
  group?: 'style' | 'content';
  description?: string;
  displayName?: string;
  defaultValue?: string | boolean | number | Record<any, any>;
}
// ...
export type CompositionVariableValueType = string | boolean | number | Record<any, any> | undefined;
type CompositionComponentPropType =
  | 'BoundValue'
  | 'UnboundValue'
  | 'DesignValue'
  | 'ComponentValue';

export type CompositionComponentPropValue<
  T extends CompositionComponentPropType = CompositionComponentPropType
> = T extends 'DesignValue'
  ? // The keys in valuesByBreakpoint are the breakpoint ids
    { type: T; valuesByBreakpoint: Record<string, CompositionVariableValueType> }
  : T extends 'BoundValue'
  ? { type: T; path: string }
  : { type: T; key: string };
// ...
export type CompositionNode = {
  definitionId: string;
  children: Array<CompositionNode>;
  variables: Record<string, CompositionComponentPropValue>;
};

export type CompositionDataSource = Record<string, DataSourceEntryValueType>;
export type CompositionUnboundValues = Record<string, { value: CompositionVariableValueType }>;

export type Breakpoint = {
  id: string;
  query: string;
  displayName: string;
  previewSize: string;
};

export type SchemaVersions = '2023-09-28' | '2023-06-27' | '2023-07-26' | '2023-08-23';

export type ExperienceComponentSettings = {
  variableDefinitions: Record<
    string,
    Omit<ComponentDefinitionVariableBase<ComponentDefinitionVariableType>, 'defaultValue'> & {
      defaultValue: CompositionComponentPropValue<'BoundValue' | 'UnboundValue'>;
    }
  >;
};

export type Composition = {
  title: string;
  slug: string;
  componentTree: {
    breakpoints: Array<Breakpoint>;
    children: Array<CompositionNode>;
    schemaVersion: SchemaVersions;
  };
  dataSource: CompositionDataSource;
  unboundValues: CompositionUnboundValues;
  usedComponents?: Array<Link<'Entry'> | ExperienceEntry>;
  componentSettings?: ExperienceComponentSettings;
};

export type ExperienceEntry = {
  sys: Entry['sys'];
  fields: Composition;
  metadata: Entry['metadata'];
};